pyproject.toml
[tool.poetry]
name = "csgogamble"
version = "1.0"
description = ""
authors = ["Jakob Endler <jakob@endler.tech>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
bs4 = "^0.0.1"
sqlite3-api = "^2.0.4"
pandas = "^1.4.2"
http-request-randomizer = "^1.3.2"
cloudscraper = "^1.2.60"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
